requirements.txt
python-nmap==0.7.1
colorama==0.4.6
psutil==5.9.8
netifaces==0.11.0
Pillow>=10.2.0
stegano>=0.10.2
cryptography>=41.0.0
